add: crated endpoint to get back all parties for a given user with type DM
